+++ packages/client/src/hooks/auth/useAuth.ts
