+++ packages/client/src/hooks/trade/useCheckCoin.ts
